calculator.cpp

+++ calculator.cpp
@@ -4,19 +4,19 @@
 
 using namespace std;
 
-float addition(int num1, int num2) {
+float addition(float num1, float num2) {
 	return(num1+num2);
 }
 
-float subtraction(int num1, int num2) {
+float subtraction(float num1, float num2) {
 	return(num1-num2);
 }
 
-float division(int num1, int num2) {
+float division(float num1, float num2) {
 	return(num1/num2);
 }
 
-float multiplication(int num1, int num2) {
+float multiplication(float num1, float num2) {
 	return(num1*num2);
 }
